sol for web3athon  2023
This project's implementation of the code will involve building decentralised applications (DApps) with frameworks like React or Angular and building smart contracts in Solidity. Implementing transaction verification, crowdfunding tools, and token production will be the main goals of the Solidity code. With the help of the DApp, community members will be able to interact with the hyperlocal solutions, record and verify transactions, take part in crowdfunding campaigns, and conduct transactions using the localised cryptocurrency. The DApp with Solidity smart contracts will enable secure and transparent interactions within the hyperlocal community, addressing concerns with trust and transparency and fostering local economic development.